+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.0.16
 
+* `NavigationChain` got its own optional id
+* `NavigationChain` and `NavigationNode` got `findChain` extension
+* `NavigationChain` and `NavigationNode` got `findNode` extension
+
 ## 0.0.15
 
 * `Versions`: